refactor(generic): extract `directory` to `filesystem` library
