+++ features/deployment/steps/.test/hooks.test.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { join } = require('node:path')
-const { directory } = require('@toa.io/libraries/generic')
+const { directory } = require('@toa.io/libraries/filesystem')
 const mock = require('@toa.io/libraries/mock')
 
 jest.mock('@cucumber/cucumber', () => mock.gherkin)
