+++ libraries/filesystem/package.json
@@ -1,0 +1,5 @@
+{
+  "private": true,
+  "main": "src/index.js",
+  "types": "types/index.d.ts"
+}
