+++ libraries/filesystem/src/directory/copy.js
@@ -1,0 +1,10 @@
+'use strict'
+
+const { promisify } = require('node:util')
+const exec = promisify(require('node:child_process').exec)
+
+const copy = async (source, target) => {
+  await exec(`cp -r "${source}/" "${target}"`)
+}
+
+exports.copy = copy
