+++ libraries/filesystem/src/directory/index.js
