+++ libraries/filesystem/src/directory/is.js
@@ -2,10 +2,6 @@
 
 const { stat } = require('node:fs/promises')
 
-/**
- * @param path {string}
- * @returns {Promise<boolean>}
- */
 const is = async (path) => {
   const entry = await stat(path)
