+++ libraries/filesystem/src/directory/remove.js
@@ -2,10 +2,6 @@
 
 const { rm } = require('node:fs/promises')
 
-/**
- * @param path {string}
- * @return {Promise<void>}
- */
 const remove = async (path) => rm(path, RECURSIVE)
 
 const RECURSIVE = { recursive: true }
