+++ libraries/filesystem/src/index.js
@@ -1,0 +1,5 @@
+'use strict'
+
+const directory = require('./directory')
+
+exports.directory = directory
