+++ libraries/filesystem/test/directory/nested/package.json
