+++ operations/src/deployment/deployment.js
@@ -2,7 +2,7 @@
 
 const { join } = require('node:path')
 const { writeFile: write } = require('node:fs/promises')
-const { directory: { copy } } = require('@toa.io/libraries/generic')
+const { directory: { copy } } = require('@toa.io/libraries/filesystem')
 const { dump } = require('@toa.io/libraries/yaml')
 
 const { merge, declare, describe } = require('./.deployment')
