+++ operations/src/deployment/images/composition.js
@@ -1,7 +1,8 @@
 'use strict'
 
 const { join } = require('node:path')
-const { hash, directory: { copy } } = require('@toa.io/libraries/generic')
+const { hash } = require('@toa.io/libraries/generic')
+const { directory: { copy } } = require('@toa.io/libraries/filesystem')
 
 const { Image } = require('./image')
