+++ operations/src/deployment/images/image.js
@@ -3,7 +3,8 @@
 const { join, posix } = require('node:path')
 const { readFile: read, writeFile: write } = require('node:fs/promises')
 
-const { directory, hash } = require('@toa.io/libraries/generic')
+const { hash } = require('@toa.io/libraries/generic')
+const { directory } = require('@toa.io/libraries/filesystem')
 
 /**
  * @implements {toa.operations.deployment.images.Image}
