+++ operations/src/deployment/images/service.js
@@ -3,7 +3,7 @@
 const { join, dirname } = require('node:path')
 
 const { Image } = require('./image')
-const { directory: { copy } } = require('@toa.io/libraries/generic')
+const { directory: { copy } } = require('@toa.io/libraries/filesystem')
 
 class Service extends Image {
   dockerfile = join(__dirname, 'service.Dockerfile')
