+++ runtime/formation/src/context.js
@@ -1,7 +1,8 @@
 'use strict'
 
 const { resolve } = require('node:path')
-const { convolve, directory: { glob } } = require('@toa.io/libraries/generic')
+const { convolve } = require('@toa.io/libraries/generic')
+const { directory: { glob } } = require('@toa.io/libraries/filesystem')
 const { load } = require('@toa.io/libraries/yaml')
 
 const { component } = require('./component')
